+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
 - Buffed `Extreme Speed` to work on allies within 40 feet.
+### Fixed
+- Fixed `Second Reincarnation` missing buff.
 
 ## [3.1.0] - 2022-12-16
 ### Added